+++ buildSrc/src/main/java/Versions.kt
@@ -28,6 +28,6 @@ object Versions {
 
     // test
     const val junit = "4.13.2"
-    const val robolectric = "4.8.1"
-    const val mockito = "4.6.1"
+    const val test_truth = "1.1.3"
+    const val mockk = "1.12.2"
 }